pyproject.toml
[tool.poetry]
name = "docs-to-markdown"
version = "0.1.0"
description = "Convert DOCX documents to Markdown format"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    { include = "docs_to_markdown.py" }
]

[tool.poetry.dependencies]
python = ">=3.9.0,<4.0"
python-docx = "^1.0.1"
docx2python = "^3.5.0"
mammoth = "^1.9.0"
html2text = "^2020.1.16"
tabulate = "^0.9.0"
beautifulsoup4 = "^4.12.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 